+++ src/ska_mid_dish_manager/models/dish_enums.py
@@ -166,3 +166,11 @@ class SPFRxCapabilityStates(enum.IntEnum):
     STANDBY = 2
     CONFIGURE = 3
     OPERATE = 4
+
+
+class NoiseDiodeMode(enum.IntEnum):
+    """SPFRx noise diode mode enums."""
+
+    OFF = 0
+    PERIODIC = 1
+    PSEUDO_RANDOM = 2