+++ tests/unit/interface/no_tango/test_cm_unhappy_paths.py
@@ -40,3 +40,73 @@ def test_track_load_static_off_with_invalid_input(
     )
     assert result_code == TaskStatus.REJECTED
     assert message == "Expected 2 arguments (off_xel, off_el) but got 3 arg(s)."
+
+
+@pytest.mark.unit
+def test_periodic_noise_diode_pars_with_invalid_input(
+    component_manager: DishManagerComponentManager,
+) -> None:
+    """
+    Verify behaviour of set_periodic_noise_diode_pars command using invalid input.
+
+    :param component_manager: the component manager under test
+    :param callbacks: a dictionary of mocks, passed as callbacks to
+        the command tracker under test
+    """
+    with pytest.raises(ValueError, match="Expected value of length 3 but got 2."):
+        component_manager.set_periodic_noise_diode_pars([1.0, 2.0])
+
+
+@pytest.mark.unit
+def test_periodic_noise_diode_pars_with_invalid_states(
+    component_manager: DishManagerComponentManager,
+) -> None:
+    """
+    Verify behaviour of set_periodic_noise_diode_pars command with invalid states.
+
+    :param component_manager: the component manager under test
+    :param callbacks: a dictionary of mocks, passed as callbacks to
+        the command tracker under test
+    """
+    with pytest.raises(
+        AssertionError,
+        match="Cannot write to periodicNoiseDiodePars."
+        " Device is not in STANDBY or MAINTENANCE state."
+        " Current state: UNKNOWN",
+    ):
+        component_manager.set_periodic_noise_diode_pars([1.0, 2.0, 3.0])
+
+
+@pytest.mark.unit
+def test_pseudo_random_noise_diode_pars_with_invalid_input(
+    component_manager: DishManagerComponentManager,
+) -> None:
+    """
+    Verify behaviour of set_pseudo_random_noise_diode_pars command using invalid input.
+
+    :param component_manager: the component manager under test
+    :param callbacks: a dictionary of mocks, passed as callbacks to
+        the command tracker under test
+    """
+    with pytest.raises(ValueError, match="Expected value of length 3 but got 2."):
+        component_manager.set_pseudo_random_noise_diode_pars([1.0, 2.0])
+
+
+@pytest.mark.unit
+def test_pseudo_random_noise_diode_pars_with_invalid_states(
+    component_manager: DishManagerComponentManager,
+) -> None:
+    """
+    Verify behaviour of set_periodic_noise_diode_pars command with invalid states.
+
+    :param component_manager: the component manager under test
+    :param callbacks: a dictionary of mocks, passed as callbacks to
+        the command tracker under test
+    """
+    with pytest.raises(
+        AssertionError,
+        match="Cannot write to pseudoRandomNoiseDiodePars."
+        " Device is not in STANDBY or MAINTENANCE state."
+        " Current state: UNKNOWN",
+    ):
+        component_manager.set_pseudo_random_noise_diode_pars([1.0, 2.0, 3.0])