+++ tests/unit/interface/tango/test_versioning.py
@@ -9,7 +9,7 @@ from tango.test_context import DeviceTestContext
 
 from ska_mid_dish_manager.devices.DishManagerDS import DishManager
 from ska_mid_dish_manager.models.constants import (
-    DEFAULT_DISH_MANAGER_TRL,
+    DEFAULT_DS_MANAGER_TRL,
     DEFAULT_SPFC_TRL,
     DEFAULT_SPFRX_TRL,
 )
@@ -54,7 +54,7 @@ class TestDishManagerVersioning:
             == ReleaseInfo.get_dish_manager_release_version()
         )
         assert build_state_json["ds_manager_device"]["version"] == ""
-        assert build_state_json["ds_manager_device"]["address"] == DEFAULT_DISH_MANAGER_TRL
+        assert build_state_json["ds_manager_device"]["address"] == DEFAULT_DS_MANAGER_TRL
         assert build_state_json["spfrx_device"]["version"] == ""
         assert build_state_json["spfrx_device"]["address"] == DEFAULT_SPFRX_TRL
         assert build_state_json["spfc_device"]["version"] == ""